Put shared logic in Layout
